Add macro to access stack slots past the register spill area
The DT_TRAMP_SP(n) macro has been renamed to DT_TRAMP_SP_SLOT(n) for
consistency with the new DT_STK_SP_SLOT(n) macro.  DT_STK_SP is new
and refers to the location of a "stack pointer" that holds the address
of a DT_STK_SP_SLOT.

This patch also removes the SCRATCH_* macros because they are obsolete.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com