Exclude %r0 from automatic register allocation
The %r0 register is used as a scratch register in various parts of the
code generator and should therefore not be among the registers that can
be allocated when dt_regset_alloc() is called.

This patch also offers a minor improvement in the regset debugging
output, reporting in-use registers by number rather than with a 'x'
in the register list.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>