Improve and unify dt_cg_prearith() and dt_cg_postarith()
Combine dt_cg_prearith() and dt_cg_postarith() into a single function
given how much functionality they share.  The implementation is now
evaluating the child node as an lvalue, to be used to load the original
value and also to store the new value.  Associative arrays and TLS
variables are a special case because they will yield a NULL lvalue if
the variable was never assigned to (or if 0 was assigned to it).  In
that case, we use dt_cg_store_var() to store the new value.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>