Rename subroutine helper (arg to tstring)
When the code generator sets up a call to a precompiled BPF function,
it emits a lot of boilerplate code.  The implementations of DTrace
subroutines typically make such calls.  So there is potential for
much code reuse.

The "path" helper is one such attempt.  Since its usefulness is
broader than just for a "path name" argument, rename this helper to
suggest its broader usage more clearly.

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>