test: Fix -lm tests
Test tst.InvalidModule4.sh was introduced as part of the DTrace port to
Linux.  It assumed that including a predicate and clause would invalidate
the module name, but that is not the case.  Rather, the module name was not
recognized in the legacy port simply because the named module did not exist.

This ill-constructed test was XFAIL, but that annotation was lifted in
patch 9574434662d0
"Remove @@xfail markers from tests that should pass with assoc support",
leading to some test failures.  Note that there is no connection between
this test and associative support.

Remove the test.

Add a -lm test for nonexisting names to match corresponding "Empty" tests.

Also add a -lm test to confirm that including a predicate and clause does
not invalidate the module name.

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>