test: Handle concurrent firing probes with exit()
A few tests have profile probes that emit an expected message and then
exit.  The problem with this is that it is possible for two such probes,
on different CPUs, to fire "concurrently."  DTrace says that when exit()
occurs, other actions already in progress on other CPUs complete.

Fix the tests to allow for more than one instance of the expected
message.  Note that this is a race condition.  The problem being fixed
was observed only intermittently.  In some test runs, the failure was
seen only 3% of the time with tst.ZeroProviderProbes.sh and only 0.06%
of the time with tst.ZeroFunctionProbes.sh.

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>