test: allow test options to be set to "-e"
The curse of echo caught us again.  Using printf keeps the -e, but even
then we have to doble-quote the argument because one layer of quotes
will be eaten by the eval, and printf with more args than format string
args reuses the format string over and over again: if it's %s, this
effectively squashes the args together with intervening spaces removed.

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>