Do not set NONALLOCA taint on VAR[IDX]
The logic for setting a NONALLOCA taint on an identifier for stores of
non-alloca, non-literal-0 values in the DT_OP_LBRAC handling in
dt_node_op2() is incorrect.  We are removing it with this patch.
Further analysis is necessary to determine whether it can be applied to
the underlying array variable identifier and how.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Nick Alcock <nick.alcock@oracle.com>