Assemble tuple components at predictable offsets
We concatenate tuple components in the tuple assembly area.  When a
component is a string, we place the following component immediately
after the terminating NUL byte.

We will use this same tuple-assembly code to support aggregation keys,
however, and for printing aggregation keys it would be nice for each
key to start at a predictable offset.  Make that change in dt_cg_arglist().

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>