test: check goodkey results
This test was failing due to lack of support for aggregation keys.
The test was tagged XFAIL.

Support for aggregation keys would cause the test to XPASS.
Unfortunately, the crucial probe action does not even execute since it
depends on vtimestamp, which is not yet implemented.  The XPASS would
be spurious since results were simply not being checked.

Add .r and .r.p files for a sanity check on results.

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>