Make sure disassembly distinguishes between gvars and agg vars
When the disassembler recognizes access of a variable by value,
it looks up a variable name that matches the offset and scope.
The scope of aggregation variables is also "global", however,
and so the presence of aggregations can lead to incorrect
disassembly of global variables.

Check the variable "kind" when looking up variables accessed
by value.

When the disassembler recognizes aggregation names, a similar
test for them can be introduced.

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>