test: Report unexpected results for unstable tests as xfail
Generally, tests are deterministically expected to pass or xfail.

Some tests fail intermittently, however, and are tagged "unstable".
Either pass or fail is acceptable, but one has to check manually
to see if the test is unstable.

If a test xpasses or fails, then if it is unstable, just report it
as xfail and unstable.

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>