Fix the -xctfpath option
The -xctfpath option was (along with all other options) being processed
after dtrace_open().  But the implementation of dtrace_open() included
the instantiation of providers and probes that triggers the loading of
the CTF archive.  As such, this option had no effect.

Instantiation of providers and probes has been moved to a new function:
dtrace_init().  This function is called after all command line options
have been processed, ensuring that -xctfpath sets the pathname to the
CTF archive before it is loaded.

If the CTF rchive specified with -xctfpath (or the default one) is not
valid, a core dump would result.  Nick Alcock submitted a patch that
solved 70% of that problen by properly checking the return value of
ctf_lookup_by_name() during program linking and flagging EDT_NOCTF if
there is an issue.  This has been incorporated into this patch.

In addition, link failure resulted in the constructed DIFO to be free'd
but upon termination, dtrace was walking the identifier hash and freeing
all DIFO and we ended up freeing the same DIFO twice.  This is fixed here
also.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>