Fix the -xverbose option
The -xverbose option should be equivalent to the -S option.  The
-xverbose option sets DTRACE_C_DIFV in the dt_cflags, but the compiler
only ever checked the cflags passed from the command line program for
the DTRACE_C_DIFV flag.  As a result, only -S had the intended
behaviour whereas -xverbose was essentially ignored.

Note that -xdisasm=N does not modify the -xverbose behaviour because the
-xverbose option is specifically defined to providw clause disassembly
(equivalent with -S without -xdisasm=N being specified).

When a D script is compiled that is not a library script or the builtin
ERROR probe, and the DTRACE_C_DIFV flag is set in dt_cflags, it is also
set in the cflags used by the compiler.

The support for -xcpp is also updated with this patch to set DTRACE_C_CPP
under the same circumstances.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>