Remove DT_AGG_NUM_COPIES
Originally, the DTrace port to BPF used two copies of aggregations,
to allow lockless access to aggregation data for producer and consumer.

Then DT_AGG_NUM_COPIES was introduced and set to 1 since the consumer
could not mmap BPF maps;  rather, it read data atomically using bpf()
system calls.  The use of twice as much locked memory for aggregations
was not justified.

Eliminate DT_AGG_NUM_COPIES, committing to the single-copy approach
and simplifying the code.

The dual-copy version used a latch sequence number.  It is now simply
a data counter;  importantly, a nonzero value indicates the presence
of data.

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>