Add a block of zeroes to use to initialize BPF maps
There are cases in BPF code where we want to read a contiguous block of
zeroes.  Since the block is read-only, it can be shared among all CPUs and
among multiple purposes to save BPF map space (locked memory).  Create such
a block.

The 'strtab' map is a BPF_MAP_TYPE_ARRAY map that already reserves a block
memory at the end of its value to placate the BPF verifier, so we can just
use that (and enlarge it if needed).

For code generation, we need to know the offset of this region in the
strtab.  The offset cannot be known until after compilation, so the
relocation mechanism is used to patch in its value..

The dt_cg_zerosptr() function should be called to obtain the pointer value
to the memory block of zeros.  It will be stored in the given register.

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>