test: Add results file for err.resize2.d
The err.resize2.d test PASSes.  Since this is an err.* test, that
means that the D script fails.  It fails, however, not because the
aggsize limit was excessive (which is the point of the test) but
simply because there is not yet support for aggregation keys.

Add a .r results file to ensure that if the D script fails,
it fails for the expected reason.

The real point of the test is to check the manual resizing option.
For now, however, we settle for reasonable behavior vis-a-vis
support for aggregation keys.

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>