Improve error checking and reporting for BPF map creation
Eugene Loh found that when BPF creation fails due to too big a map,
E2BIG is returned.  This got reported as "Argument list too long"
(using strerror()).  That is not very helpful.  It will now report
"Too big".

In addition, the key and value sizes are size_t entities in the main
caller (dt_bpf_gmap_create()), while create_map() accepts ints for
these values.  And the maximum number of map entries is an int in
dt_bpf_gmap_create() but is at times calculated based on size_t values.
And the bpf(BPF_MAP_CREATE) syscall expects uint32_t values for all
three.  That is now cleaned up.

The debug output in create_gmap() also got a small modification to no
longer report sizes in the success report since it is already reported
at the beginning of the function.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>