Fix records for aggregation descriptions
The implementation of aggregations erroneously registered the value of
an aggregation as a series of 64-bit values, each with their own record
description.  However, aggregations in their most generic form should
comprise the following records:

(1) variable ID
(2) keys (values of the index tuple, if any)
(3) latch sequence number
(4) aggregation data

(1) and (2) are records that identify the aggregation, whereas (3) and
(4) are data that is generated by the producer.

Change the aggregation record descriptions to match the layout described
above.

Note that the variable ID is now stored as a 32-bit value whereas DTrace
used to store it as a 64-bit value.  This change is valid because the
variable ID is an identifier ID and therefore cannot exceed the range of
a 32-bit integer,

Support for aggregation keys will be added in later patches.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>