Clear dt_aggregate each snapshot
The producer (kernel) gathers aggregation data.  The consumer
(user space) snapshots (reads) the kernel data and maintains
its own copy in dt_aggregate.

In legacy DTrace, the consumer would snapshot incremental kernel
data and either initialize or update an entry in dt_aggregate,
depending on whether the corresponding entry already existed.

In the port to BPF, the kernel data is kept in BPF maps that are
up-to-date, not incremental since the last snapshot.  So a second
mechanism was added to decide whether to initialize or update a
dt_aggregate entry:  for the first CPU, even if an entry has been
found, new data is copied (rather than aggregated) in.

When we support aggregation keys, the new mechanism will not help.
It will be possible for a key to appear on some other CPU than
the first.

Nevertheless, we cannot blindly keep re-aggregating data that
is already in dt_aggregate.

Therefore, clear aggregation data in dt_aggregate's hash entries
at each snapshot.  Use dtrace_aggregate_clear() for this purpose.
Note that this function is apparently unused in both legacy DTrace
as well as the port to BPF.  Note also that the function needs to
be corrected to account for for min() and max()

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>