test: fix test/demo/user/libc.d
It seems that /bin/date is more likely to exist than /usr/bin/date.
Also, the test needs a '@@trigger: none' marker to ensure it actually
gets executed.

The probe range is limited to functions that match m* to reduce the
resource needs for this test.  Since this includes functions like
malloc() it does not affect the operation of the test.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>