test: Bump up locked memory in lockmem tests
The default aggsize is 4M.  This is the dominant drain on locked
memory for a simple script like "BEGIN { @ = avg(1) }".

The lockmem tests check behavior when the lockmem limit is too
small; anything less than 4M is suitable.

They also check when the lockmem limit is large enough.  We were
using a lockmem limit of 16M, which is clearly inadequate even
for only a few CPUs.

Bump the "large enough" sizes up to 256M, which should be sufficient
even for dozens of CPUs.

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>