test: Speed up check for modname-symname duplicates
When we check "lookup by name", a mismatch is allowed if there are
duplicates with the same modname and symname.  So, we must be able
to detect when such duplicates occur.  We were using a linear search.

Add a quick sort by modname/symname and then initialize a field
"is_duplicate" to speed up detection of duplicates.

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>